DB functionalities added and called in thier respective classes

+++ app/src/main/java/pk/edu/pucit/kitchen_witchenmc_project/viewAdapter/DishRViewAdapter.java
@@ -42,7 +42,7 @@ public class DishRViewAdapter extends RecyclerView.Adapter<DishRViewAdapter.View
     @Override
     public void onBindViewHolder(@NonNull ViewHolder holder, final int position) {
         holder.tv_dname.setText(dish_items.get(position).getName());
-        holder.tv_price.setText(Integer.toString(dish_items.get(position).getPrice()));
+        holder.tv_price.setText(Long.toString(dish_items.get(position).getPrice()));
         Picasso.get().load(dish_items.get(position).getImg()).into(holder.tv_dicon);
         holder.tv_dbtn.setOnClickListener(new View.OnClickListener() {
             @Override
